back_end/llamaIndex_app/pyproject.toml
[project]
name = "llamaindex_app"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "llama-index>=0.14.12",
    "llama-index-embeddings-huggingface>=0.6.1",
    "llama-index-embeddings-ollama>=0.8.6",
    "llama-index-llms-ollama>=0.9.1",
    "llama-index-readers-file>=0.5.6",
    "llama-index-tools-metaphor>=0.4.1",
    "llama-index-vector-stores-milvus>=0.9.6",
    "pymilvus[milvus-lite]>=2.6.6",
    "pymupdf>=1.26.7",
    "python-dotenv"
]


[tool.uv]
package = true

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]